pyproject.toml
[project]
name = "sc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "fastapi>=0.118.0",
    "greenlet>=3.2.4",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]
